[TRTLLM-13231][chore] Move the top-p-decay fused ops into ops.vanilla.Fusions
They are pure-torch compiled ops; Fusions in ops/vanilla.py is the existing
home for torch.compile'd fused helpers (same impl + mark_dynamic wrapper
pattern as gather_log_softmax), so the standalone ops/top_p_decay.py module
is folded into it and call sites use Fusions.top_p_decay_{gather,update}.
